requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pillow
imagehash
requests
opencv-python-headless
numpy
